requirements.txt
hydra-core~=1.1.0
numpy~=1.19.5
omegaconf~=2.1.1
pandas~=1.1.5
pytorch-lightning~=1.4.1
tensorboard~=2.4.0
torchmetrics~=0.5.1
tqdm~=4.62.2
torch~=1.10.2+cu113